assignment_3/requirements.txt
openai
groq
cohere
google-generativeai
requests
python-dotenv
streamlit